# Optimize regexp matcher with list of alternations ending with .*