@pracucci 2023-09-29
> LGTM. This took me a while to get my head around this.

This stuff is very tricky 🤞